requirements.txt
Django==4.1.1
psycopg-binary==3.1.2
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1
djangorestframework-simplejwt==5.2.0
whitenoise==6.2.0
django-tenants==3.4.5
psycopg2==2.9.3
